requirements.txt
gradio
ultralytics
opencv-python
torch
torchvision
pytube
mss
Pillow
